packages/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.25.0
openai==1.3.5
qdrant-client==2.6.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
httpx==0.25.2
